requirements.txt
django>=4.2,<4.3
djangorestframework
markdown
django-filter
drf-spectacular
django-rest-knox
django-cors-headers
python-decouple
dj-rest-auth[with_social]
djangorestframework-simplejwt